# api/openapi.yaml
openapi: 3.0.3
info:
  title: JuanNiang-RAG-Service API
  description: |
    scoop（分库）+ tag（uuid）↔ 向量 的存储与检索服务，供主 Agent 通过 HTTP 调用。

    - scoop 白名单：`knowledge` / `memory` / `groupmgr` / `plugin`；检索只在目标分库内进行
    - 原始文档与 uuid 由 Agent 保管，本服务只做向量化、检索与删除
    - 长文本服务端自动分块，Agent 契约始终为 tag ↔ 全文
    - 同一 tag 全局唯一归属一个分库，跨分库写入/删除 → 409
    - 查询与写入互不阻塞；写接口返回即持久化完成（读己之写）
    - 分数 `score` 为相似度（0~1），越高越相似
  version: 2.0.0
  contact:
    name: JuanNiang-Neo
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:3000
    description: 本地默认部署

# 公开服务，无认证
security: []

tags:
  - name: scoops
    description: scoop 分库内 tag ↔ 向量 的写入与检索
  - name: ops
    description: 运维与诊断

paths:
  /scoops/{scoop}/tags/{tag}:
    parameters:
      - $ref: '#/components/parameters/ScoopPath'
      - $ref: '#/components/parameters/TagPath'
    put:
      tags: [scoops]
      summary: 更新（upsert）
      description: |
        tag 已存在则覆写其全部向量，不存在则新建（写入 scoop 分库）。
        长文（>512 token）服务端自动分块，全部块向量挂同一 tag；
        响应返回时数据已持久化并发布，之后检索立即可见。
        同一 tag 已归属其他分库时返回 409。
      operationId: upsertTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertRequest'
      responses:
        '200':
          description: 入库成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpsertResponse'
              example:
                scoop: knowledge
                tag: 3af2b489-b13a-42e4-af98-fe89d0e6b00e
                chunk_count: 4
                truncated: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [scoops]
      summary: 删除
      description: 删除 tag 及其全部块向量（只能删除属于该 scoop 的 tag）。
      operationId: deleteTag
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: string
                    format: uuid
                    description: 被删除的 tag
              example:
                deleted: 3af2b489-b13a-42e4-af98-fe89d0e6b00e
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /scoops/{scoop}/tags:
    parameters:
      - $ref: '#/components/parameters/ScoopPath'
    get:
      tags: [scoops]
      summary: 列出分库内全部 tag
      description: 分页返回该 scoop 内的 tag 列表（含块数），供对账/清理使用。
      operationId: listTags
      parameters:
        - name: page
          in: query
          required: false
          description: 页码（从 1 起，默认 1）
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          required: false
          description: 每页条数（默认 20，上限 100，超出自动钳制）
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: tag 列表（按插入顺序）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagListResponse'
              example:
                total: 128
                page: 1
                page_size: 20
                items:
                  - tag: 3af2b489-b13a-42e4-af98-fe89d0e6b00e
                    chunk_count: 4
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /scoops/{scoop}/tags/batch-delete:
    post:
      tags: [scoops]
      summary: 批量删除
      description: |
        批量删除 tag（全部属于该 scoop）。逐条返回结果：不存在/归属其他分库
        的 tag 标记为 deleted=false 并携带错误，不影响其他条目。单次上限 500 条。
      operationId: batchDelete
      parameters:
        - $ref: '#/components/parameters/ScoopPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDeleteRequest'
            example:
              tags:
                - 11111111-1111-1111-1111-111111111111
                - 22222222-2222-2222-2222-222222222222
      responses:
        '200':
          description: 批量删除完成（逐条结果，顺序与请求一致）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDeleteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /metrics:
    get:
      tags: [ops]
      summary: Prometheus 指标
      description: |
        文本格式（Prometheus exposition format v0.0.4）。
        指标前缀 rag_，gauge（规模/就绪）在 scrape 时实时读取。
      operationId: getMetrics
      responses:
        '200':
          description: 指标文本
          content:
            text/plain:
              schema:
                type: string
              example: |
                # HELP rag_embedder_ready 嵌入模型是否就绪（0/1）
                # TYPE rag_embedder_ready gauge
                rag_embedder_ready 1
        '500':
          $ref: '#/components/responses/InternalError'

  /scoops/{scoop}/tags/batch:
    post:
      tags: [scoops]
      summary: 批量更新
      description: |
        多条文本一次批量 upsert：一次嵌入推理 + 一次发布 + 一次持久化，
        批量成本远低于逐条调用。全部条目必须隶属同一个 scoop。
        逐条返回结果，单条失败（如归属冲突）不影响其他条目。
      operationId: batchUpsert
      parameters:
        - $ref: '#/components/parameters/ScoopPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchUpsertRequest'
            example:
              items:
                - tag: 11111111-1111-1111-1111-111111111111
                  text: 第一条文本
                - tag: 22222222-2222-2222-2222-222222222222
                  text: 第二条文本
      responses:
        '200':
          description: 批量处理完成（含逐条结果，顺序与请求一致）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /scoops/{scoop}/tags/search:
    get:
      tags: [scoops]
      summary: 检索
      description: |
        输入文本，返回按相似度排序的 tag 列表；只在指定 scoop 分库内检索，
        其他分库的向量不参与。
        块级命中去重聚合（每 tag 取最高分）；查询文本自动加 bge 指令前缀；
        相同 `q` 短时间重复查询命中 LRU 缓存（免推理）。
      operationId: searchTags
      parameters:
        - $ref: '#/components/parameters/ScoopPath'
        - name: q
          in: query
          required: true
          description: 查询文本（不能为空）
          schema:
            type: string
        - name: k
          in: query
          required: false
          description: 返回条数（范围 1–100，超出自动钳制）
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: min_score
          in: query
          required: false
          description: 相似度下限过滤（0~1），低于该值的 tag 不返回
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 1
      responses:
        '200':
          description: 检索结果（按相似度降序）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              example:
                results:
                  - tag: 3af2b489-b13a-42e4-af98-fe89d0e6b00e
                    score: 0.888
                  - tag: 22222222-2222-2222-2222-222222222222
                    score: 0.698
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /health:
    get:
      tags: [ops]
      summary: 健康检查
      operationId: getHealth
      responses:
        '200':
          description: 服务可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
                scoops:
                  knowledge: { tags: 128, chunks: 340 }
                  memory: { tags: 56, chunks: 60 }
                  groupmgr: { tags: 2300, chunks: 2300 }
                  plugin: { tags: 0, chunks: 0 }

  /info:
    get:
      tags: [ops]
      summary: 服务信息
      description: |
        embedding 模型状态 + 进程内存 + 各分库规模。
        `model.ready=false` 时查看 `model.error` 获取加载失败原因；
        `memory` 仅 Linux（读 /proc/self/status），其他平台为 null。
      operationId: getInfo
      responses:
        '200':
          description: 服务信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
              example:
                status: ok
                model:
                  ready: true
                  model_name: bge-small-zh-v1.5
                  dim: 512
                  n_params: 23691264
                  n_threads: 4
                  n_ctx: 4096
                  error: null
                memory:
                  rss_kb: 81244
                  vsize_kb: 1915772
                scoops:
                  knowledge: { tags: 128, chunks: 340 }
                  memory: { tags: 56, chunks: 60 }
                  groupmgr: { tags: 2300, chunks: 2300 }
                  plugin: { tags: 0, chunks: 0 }

components:
  parameters:
    ScoopPath:
      name: scoop
      in: path
      required: true
      description: 分库名（白名单枚举，非法值 → 400）
      schema:
        type: string
        enum: [knowledge, memory, groupmgr, plugin]
    TagPath:
      name: tag
      in: path
      required: true
      description: tag（uuid），Agent 侧文档标识
      schema:
        type: string
        format: uuid

  schemas:
    UpsertRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: 要入库的文本内容（空文本返回 400）
      example:
        text: Rust 是一种系统编程语言，强调内存安全与零成本抽象。

    UpsertResponse:
      type: object
      required: [scoop, tag, chunk_count, truncated]
      properties:
        scoop:
          type: string
          enum: [knowledge, memory, groupmgr, plugin]
          description: 入库的分库（原样返回）
        tag:
          type: string
          format: uuid
          description: 入库的 tag（原样返回）
        chunk_count:
          type: integer
          minimum: 1
          description: 实际入库块数。≤512 token 为 1；长文自动分块后 >1
        truncated:
          type: boolean
          description: 是否存在被截断的块（单块超 512 token 时截断处理）

    BatchUpsertRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/BatchItem'

    BatchItem:
      type: object
      required: [tag, text]
      properties:
        tag:
          type: string
          format: uuid
        text:
          type: string

    BatchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          description: 逐条结果，顺序与请求一致
          items:
            $ref: '#/components/schemas/BatchItemResponse'

    BatchItemResponse:
      type: object
      required: [tag, chunk_count, truncated, error]
      properties:
        tag:
          type: string
          format: uuid
        chunk_count:
          type: integer
        truncated:
          type: boolean
        error:
          type: string
          nullable: true
          description: null 表示成功；否则为该条目失败原因（如归属冲突），该条目未入库

    SearchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchHit'

    SearchHit:
      type: object
      required: [tag, score]
      properties:
        tag:
          type: string
          format: uuid
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: 相似度（0~1），越高越相似

    TagListResponse:
      type: object
      required: [total, page, page_size, items]
      properties:
        total:
          type: integer
          description: 该分库 tag 总数
        page:
          type: integer
          description: 当前页码（从 1 起）
        page_size:
          type: integer
          description: 每页条数
        items:
          type: array
          items:
            $ref: '#/components/schemas/TagListItem'

    TagListItem:
      type: object
      required: [tag, chunk_count]
      properties:
        tag:
          type: string
          format: uuid
        chunk_count:
          type: integer
          minimum: 1
          description: 该 tag 的块向量数（长文分块后 >1）

    BatchDeleteRequest:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          minItems: 1
          maxItems: 500
          items:
            type: string
            format: uuid

    BatchDeleteResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          description: 逐条结果，顺序与请求一致
          items:
            $ref: '#/components/schemas/BatchDeleteItemResponse'

    BatchDeleteItemResponse:
      type: object
      required: [tag, deleted, error]
      properties:
        tag:
          type: string
          format: uuid
        deleted:
          type: boolean
          description: 是否成功删除
        error:
          type: string
          nullable: true
          description: null 表示成功；否则为该条失败原因（如不存在、归属其他分库）

    ScoopStats:
      type: object
      required: [tags, chunks]
      properties:
        tags:
          type: integer
          description: 该分库 tag 总数
        chunks:
          type: integer
          description: 该分库块向量总数（长文分块后 chunks ≥ tags）

    ScoopStatsMap:
      type: object
      description: 各分库规模统计（knowledge / memory / groupmgr / plugin）
      additionalProperties:
        $ref: '#/components/schemas/ScoopStats'

    HealthResponse:
      type: object
      required: [status, scoops]
      properties:
        status:
          type: string
          enum: [ok]
        scoops:
          $ref: '#/components/schemas/ScoopStatsMap'

    InfoResponse:
      type: object
      required: [status, model, memory, scoops]
      properties:
        status:
          type: string
          enum: [ok]
        model:
          $ref: '#/components/schemas/EmbedderInfo'
        memory:
          type: object
          nullable: true
          allOf:
            - $ref: '#/components/schemas/MemoryInfo'
          description: 非 Linux 平台为 null
        scoops:
          $ref: '#/components/schemas/ScoopStatsMap'

    EmbedderInfo:
      type: object
      required: [ready, model_name, dim, n_params, n_threads, n_ctx, error]
      properties:
        ready:
          type: boolean
          description: 模型是否加载成功、可服务
        model_name:
          type: string
          nullable: true
          description: GGUF 元数据里的模型名
        dim:
          type: integer
          nullable: true
          description: 嵌入维度（bge-small-zh-v1.5 = 512）
        n_params:
          type: integer
          nullable: true
          format: int64
          description: 模型参数量
        n_threads:
          type: integer
          nullable: true
          description: 推理线程数
        n_ctx:
          type: integer
          nullable: true
          description: 实际上下文长度（llama.cpp 可能因多序列配置向上取整）
        error:
          type: string
          nullable: true
          description: null 表示正常；否则为加载/初始化失败原因

    MemoryInfo:
      type: object
      required: [rss_kb, vsize_kb]
      properties:
        rss_kb:
          type: integer
          format: int64
          description: 进程常驻物理内存（kB）
        vsize_kb:
          type: integer
          format: int64
          description: 进程虚拟内存（kB）

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 错误描述

  responses:
    BadRequest:
      description: 参数错误（text/q 为空、items 为空、未知 scoop 等）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: '未知分库: foo（白名单: knowledge, memory, groupmgr, plugin）'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'tag 不存在: 3af2b489-b13a-42e4-af98-fe89d0e6b00e'
    Conflict:
      description: tag 归属冲突（同一 tag 已归属其他分库，跨分库写入/删除被拒绝）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'tag 3af2b489-b13a-42e4-af98-fe89d0e6b00e 已归属分库 knowledge，跨分库重复写入被拒绝'
    InternalError:
      description: 服务内部错误（模型未加载、存储损坏、嵌入失败等）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: '嵌入服务错误: 模型加载失败: ...'